fix: WGSL runtime errors across 35 variation files
Fixed 5 categories of WGSL shader compilation errors:

1. Ternary operators in use-gpu bodies replaced with std.select()
   WGSL does not support JS ternary for runtime values
   17 files: cschVar, cschqVar, ediscVar, secant2Var, wdiscVar,
   postCrop, hole2Var, sierCarpetVar, dragonVar, hilbertVar, ejuliaVar,
   onionVar, onion2Var, barycentroidVar, blockYVar, boarders2Var,
   cutFractalVar, cutApollonianVar, blur/index

2. Uninitialized let declarations - added initializers
   10 files: truchet2Var, modulusVar, holeVar, collideoscopeVar,
   whorlVar, sattractorVar, cellVar, circlize2Var, circlizeVar, hole2Var

3. Return pos replaced with vec2f(pos.x, pos.y)
   2 files: fourthVar, bwraps7Var

4. Shift operators changed to u32 (n << u32(13))
   3 files: circleLinearVar, circleRandVar, pixelFlowVar

5. Fixed pixelFlowVar param mismatches in example15.ts and utils.ts

Verified with pnpm check (typecheck + lint + fmt + validate-wgsl).

+++ packages/app/src/flame/variations/parametric/general/circleRandVar.tsx
@@ -1,4 +1,4 @@
-import { f32, i32, struct, vec2f } from 'typegpu/data'
+import { f32, i32, struct, u32, vec2f } from 'typegpu/data'
 import { floor, select, sqrt } from 'typegpu/std'
 import { RangeEditor } from '@/components/Sliders/ParametricEditors/RangeEditor'
 import { editorProps } from '@/components/Sliders/ParametricEditors/types'
@@ -78,12 +78,12 @@ export const circleRandVar = parametricVariation(
     const U = sqrt(rX * rX + rY * rY)
 
     const n1 = i32(M + P.seed) + i32(N) * 57
-    const n1x = (n1 << 13) ^ n1
+    const n1x = (n1 << u32(13)) ^ n1
     const z1 =
       f32((n1x * (n1x * n1x * 15731 + 789221) + 1376312589) & 0x7fffffff) * AM
 
     const n2 = i32(M + 10.0) + i32(N + 3.0) * 57
-    const n2x = (n2 << 13) ^ n2
+    const n2x = (n2 << u32(13)) ^ n2
     const V =
       (0.3 +
         0.7 *
